ai-recommendation-engine/requirements.txt
sentence-transformers>=3.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.10.0
scikit-learn>=1.6.0
numpy>=2.1.0
python-multipart>=0.0.17
PyPDF2>=3.0.0
python-docx>=1.1.0
python-pptx>=1.0.0
google-genai
